third_party/WebKit/LayoutTests/imported/wpt/html/webappapis/scripting/processing-model-2/compile-error-in-body-onerror-expected.txt

CONSOLE ERROR: line 19: Uncaught SyntaxError: Unexpected token )
CONSOLE ERROR: line 17: Uncaught SyntaxError: Unexpected token )
This is a testharness.js-based test.
FAIL window.onerror - compile error in <body onerror> assert_false: ran expected false got true
Harness: the test ran to completion.
